+++ tests/utils/password.value-object.spec.ts
@@ -76,7 +76,8 @@ describe('password.value-object', () => {
 
 		it('should password to be equal', () => {
 			const passA = PasswordValueObject.create('123456abc!').value();
-			const passB = PasswordValueObject.create('123456abc!').value();
+			/**  @todo: fix inference type on clone to get instance type */
+			const passB = passA.clone() as PasswordValueObject;
 			const isEqual = passA.isEqual(passB);
 			expect(isEqual).toBe(true);
 		});